fix: resolve electron.remote error under Local 9
@electron/remote needs to be a direct dependency and not a
devDependency under Local 9 (which upgraded Electron).
Otherwise, the add-on fails to load and this error is visible
in Local's console:

 /Applications/Local Beta.app/Contents/Resources/app.asar/node_modules/
 @sentry/utils/dist/instrument.js:111 Error Loading Add-on: /Users/
 user.name/Library/Application Support/Local Beta/addons/
 local-addon-image-optimizer/lib/renderer.js
 TypeError: i.isDesktopCapturerEnabled is not a function

An alternative is to refactor to replace @electron/remote with
IPC calls, but that is a larger lift with bigger testing impact given
use of remote.Menu, remote.MenuItem, remote.shell, and the
reportAnalytics calls that currently trigger on menu events.